[JEWEL-1258] Add missing remember keys in Markdown renderers
