+++ platform/jewel/markdown/core/intellij.platform.jewel.markdown.core.iml
@@ -58,6 +58,7 @@
     <orderEntry type="module" module-name="intellij.libraries.junit4" scope="TEST" />
     <orderEntry type="module" module-name="intellij.libraries.compose.runtime.desktop" />
     <orderEntry type="module" module-name="intellij.libraries.compose.foundation.desktop.junit" scope="TEST" />
+    <orderEntry type="module" module-name="intellij.platform.jewel.markdown.testing" scope="TEST" />
     <orderEntry type="module" module-name="intellij.libraries.jsoup" />
   </component>
 </module>
