+++ platform/jewel/markdown/extensions/gfm-tables/BUILD.bazel
@@ -43,20 +43,32 @@ jvm_library(
 
 jvm_library(
   name = "gfm-tables_test_lib",
-  module_name = "intellij.platform.jewel.markdown.extensions.gfmTables",
   visibility = ["//visibility:public"],
-  srcs = glob([], allow_empty = True),
+  srcs = glob(["src/test/kotlin/**/*.kt", "src/test/kotlin/**/*.java", "src/test/kotlin/**/*.form"], allow_empty = True),
   kotlinc_opts = ":custom_gfm-tables",
-  runtime_deps = [
-    ":gfm-tables",
+  associates = [":gfm-tables"],
+  deps = [
+    "@lib//:kotlin-stdlib",
+    "//libraries/kotlinx/coroutines/core",
     "//libraries/kotlinx/coroutines/core:core_test_lib",
+    "@lib//:jetbrains-annotations",
+    "//platform/jewel/markdown/core",
     "//platform/jewel/markdown/core:core_test_lib",
+    "//platform/jewel/ui",
     "//platform/jewel/ui:ui_test_lib",
+    "//platform/jewel/foundation",
     "//platform/jewel/foundation:foundation_test_lib",
+    "//libraries/compose-foundation-desktop",
     "//libraries/compose-foundation-desktop:compose-foundation-desktop_test_lib",
+    "//libraries/compose-runtime-desktop",
     "//libraries/compose-runtime-desktop:compose-runtime-desktop_test_lib",
+    "@lib//:platform-jewel-markdown-extensions-gfm_tables-commonmark-ext-gfm-tables",
+    "//libraries/junit4",
+    "//libraries/junit4:junit4_test_lib",
     "//libraries/compose-foundation-desktop-junit",
     "//libraries/compose-foundation-desktop-junit:compose-foundation-desktop-junit_test_lib",
+    "//platform/jewel/markdown/testing",
+    "//platform/jewel/markdown/testing:testing_test_lib",
   ],
   plugins = ["@lib//:compose-plugin"]
 )
@@ -67,3 +79,12 @@ exports_files([
   "intellij.platform.jewel.markdown.extensions.gfmTables.iml",
 ], visibility = ["//visibility:public"])
 ### auto-generated section `iml intellij.platform.jewel.markdown.extensions.gfmTables` end
+
+### auto-generated section `test intellij.platform.jewel.markdown.extensions.gfmTables` start
+load("@community//build:tests-options.bzl", "jps_test")
+
+jps_test(
+  name = "gfm-tables_test",
+  runtime_deps = [":gfm-tables_test_lib"]
+)
+### auto-generated section `test intellij.platform.jewel.markdown.extensions.gfmTables` end
