+++ platform/jewel/markdown/testing/BUILD.bazel
@@ -1,0 +1,56 @@
+### auto-generated section `build intellij.platform.jewel.markdown.testing` start
+load("//build:compiler-options.bzl", "create_kotlinc_options")
+load("@rules_jvm//:jvm.bzl", "jvm_library")
+
+create_kotlinc_options(
+  name = "custom_testing",
+  opt_in = [
+    "androidx.compose.ui.ExperimentalComposeUiApi",
+    "androidx.compose.foundation.ExperimentalFoundationApi",
+    "org.jetbrains.jewel.foundation.ExperimentalJewelApi",
+    "org.jetbrains.jewel.foundation.InternalJewelApi",
+  ],
+  x_context_parameters = True
+)
+
+jvm_library(
+  name = "testing",
+  module_name = "intellij.platform.jewel.markdown.testing",
+  visibility = ["//visibility:public"],
+  srcs = glob(["src/main/kotlin/**/*.kt", "src/main/kotlin/**/*.java", "src/main/kotlin/**/*.form"], allow_empty = True),
+  kotlinc_opts = ":custom_testing",
+  deps = [
+    "@lib//:kotlin-stdlib",
+    "@lib//:jetbrains-annotations",
+    "//platform/jewel/markdown/core",
+    "//platform/jewel/ui",
+    "//platform/jewel/foundation",
+    "//libraries/compose-foundation-desktop",
+    "//libraries/compose-runtime-desktop",
+  ],
+  plugins = ["@lib//:compose-plugin"]
+)
+
+jvm_library(
+  name = "testing_test_lib",
+  module_name = "intellij.platform.jewel.markdown.testing",
+  visibility = ["//visibility:public"],
+  srcs = glob([], allow_empty = True),
+  kotlinc_opts = ":custom_testing",
+  runtime_deps = [
+    ":testing",
+    "//platform/jewel/markdown/core:core_test_lib",
+    "//platform/jewel/ui:ui_test_lib",
+    "//platform/jewel/foundation:foundation_test_lib",
+    "//libraries/compose-foundation-desktop:compose-foundation-desktop_test_lib",
+    "//libraries/compose-runtime-desktop:compose-runtime-desktop_test_lib",
+  ],
+  plugins = ["@lib//:compose-plugin"]
+)
+### auto-generated section `build intellij.platform.jewel.markdown.testing` end
+
+### auto-generated section `iml intellij.platform.jewel.markdown.testing` start
+exports_files([
+  "intellij.platform.jewel.markdown.testing.iml",
+], visibility = ["//visibility:public"])
+### auto-generated section `iml intellij.platform.jewel.markdown.testing` end
