+++ platform/jewel/markdown/testing/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    jewel
+    alias(libs.plugins.composeDesktop)
+    alias(libs.plugins.compose.compiler)
+}
+
+dependencies {
+    api(projects.markdown.core)
+}
