+++ platform/jewel/markdown/testing/module-content.yaml
@@ -1,0 +1,3 @@
+- name: dist.all/lib/intellij.platform.jewel.markdown.testing.jar
+  modules:
+  - name: intellij.platform.jewel.markdown.testing
